# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"LABELGUARD","description":"AI-Assisted Legal Metrology Compliance and Inspection Platform","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/health":{"get":{"tags":["Health","Health"],"summary":"Health Check","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["Authentication"],"summary":"Get Me","operationId":"get_me_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CurrentUserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/auth/admin-test":{"get":{"tags":["Authentication"],"summary":"Admin Test","operationId":"admin_test_api_v1_auth_admin_test_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/inspections":{"get":{"tags":["Inspections"],"summary":"Get Inspections","operationId":"get_inspections_api_v1_inspections_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/InspectionResponse"},"type":"array","title":"Response Get Inspections Api V1 Inspections Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["Inspections"],"summary":"Create New Inspection","operationId":"create_new_inspection_api_v1_inspections_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/inspections/{inspection_id}":{"get":{"tags":["Inspections"],"summary":"Get Single Inspection","operationId":"get_single_inspection_api_v1_inspections__inspection_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"inspection_id","in":"path","required":true,"schema":{"type":"integer","title":"Inspection Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/inspections/{inspection_id}/images":{"post":{"tags":["Inspection Images"],"summary":"Upload Inspection Image","operationId":"upload_inspection_image_api_v1_inspections__inspection_id__images_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"inspection_id","in":"path","required":true,"schema":{"type":"integer","title":"Inspection Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_inspection_image_api_v1_inspections__inspection_id__images_post"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Inspection Images"],"summary":"Get Inspection Images","operationId":"get_inspection_images_api_v1_inspections__inspection_id__images_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"inspection_id","in":"path","required":true,"schema":{"type":"integer","title":"Inspection Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ImageResponse"},"title":"Response Get Inspection Images Api V1 Inspections  Inspection Id  Images Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_inspection_image_api_v1_inspections__inspection_id__images_post":{"properties":{"image_type":{"type":"string","title":"Image Type"},"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["image_type","file"],"title":"Body_upload_inspection_image_api_v1_inspections__inspection_id__images_post"},"ComplianceStatus":{"type":"string","enum":["pending","compliant","non_compliant","review"],"title":"ComplianceStatus"},"CurrentUserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"full_name":{"type":"string","title":"Full Name"},"email":{"type":"string","title":"Email"},"role":{"type":"string","title":"Role"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["id","full_name","email","role","is_active"],"title":"CurrentUserResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ImageResponse":{"properties":{"id":{"type":"integer","title":"Id"},"inspection_id":{"type":"integer","title":"Inspection Id"},"file_name":{"type":"string","title":"File Name"},"file_path":{"type":"string","title":"File Path"},"image_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Type"},"file_size":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"File Size"},"mime_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mime Type"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","inspection_id","file_name","file_path","image_type","file_size","mime_type","created_at"],"title":"ImageResponse"},"InspectionCreate":{"properties":{"product_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Product Id"}},"type":"object","title":"InspectionCreate"},"InspectionResponse":{"properties":{"id":{"type":"integer","title":"Id"},"inspector_id":{"type":"integer","title":"Inspector Id"},"product_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Product Id"},"reference_number":{"type":"string","title":"Reference Number"},"status":{"$ref":"#/components/schemas/InspectionStatus"},"compliance_status":{"$ref":"#/components/schemas/ComplianceStatus"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"completed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completed At"}},"type":"object","required":["id","inspector_id","product_id","reference_number","status","compliance_status","created_at","completed_at"],"title":"InspectionResponse"},"InspectionStatus":{"type":"string","enum":["created","processing","completed","failed"],"title":"InspectionStatus"},"LoginRequest":{"properties":{"email":{"type":"string","maxLength":255,"minLength":3,"title":"Email"},"password":{"type":"string","maxLength":128,"minLength":8,"title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}